requirements.txt
# Core Flask
Flask==3.0.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1

# ORM + Database
SQLAlchemy==2.0.30
# Forms & Validation
WTForms==3.1.2
email-validator==2.1.0.post1

# Environment variables
python-dotenv==1.0.1

# Payment API
stripe==9.7.0

# QR + Ticket Generation
qrcode==7.4.2
reportlab==4.2.2

# Utility + Security
Werkzeug==3.0.3

# Gunicorn for Heroku
gunicorn==21.2.0


Flask==3.0.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
WTForms==3.1.2
email-validator==2.1.0.post1
python-dotenv==1.0.1
stripe==9.7.0
qrcode==7.4.2
reportlab==4.2.2
Werkzeug==3.0.3
gunicorn==21.2.0
boto3==1.35.27
pandas==2.2.3